pyproject.toml
[tool.poetry]
name = "python-challenge"
version = "0.1.0"
description = ""
authors = ["João Corrêa <jfrcorrea@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
boto3 = "1.35.19"
localstack = "3.7.2"
aws-cdk-lib = "2.158.0"
pydantic = "2.9.1"
awscli-local = "0.22.0"
awscli = "1.34.19"


[tool.poetry.group.dev.dependencies]
flake8-docstrings = "1.7.0"
isort = "5.13.2"
black = "24.8.0"
pytest = "8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 121

[tool.isort]
profile = "black"
